feat: remove dummifier
