+++ whois-scheduler/src/test/java/net/ripe/db/whois/scheduler/task/export/DecorationStrategyTest.java
@@ -1,13 +1,12 @@
 package net.ripe.db.whois.scheduler.task.export;
-import net.ripe.db.whois.common.rpsl.DummifierCurrent;
+
 import net.ripe.db.whois.common.rpsl.DummifierNrtm;
 import net.ripe.db.whois.common.rpsl.RpslObject;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -15,14 +14,12 @@ import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.nullValue;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 public class DecorationStrategyTest {
     RpslObject object;
     @Mock
     DummifierNrtm dummifier;
-    @Mock DummifierCurrent dummifierCurrent;
 
     @BeforeEach
     public void setUp() throws Exception {
@@ -67,18 +64,4 @@ public class DecorationStrategyTest {
         verify(dummifier, never()).dummify(3, object);
     }
 
-    @Test
-    public void decorate_dummify_proposed_allowed() {
-        DecorationStrategy subject = new DecorationStrategy.DummifyCurrent(dummifierCurrent);
-        final RpslObject object = RpslObject.parse("role: Test Role\nnic-hdl: TR1-TEST");
-
-        when(dummifierCurrent.isAllowed(3, object)).thenReturn(true);
-        when(dummifierCurrent.dummify(3, object)).thenReturn(object);
-
-        final RpslObject decorated = subject.decorate(object);
-
-        assertThat(object, is(decorated));
-        verify(dummifierCurrent).isAllowed(3, object);
-        verify(dummifierCurrent).dummify(3, object);
-    }
 }
